% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fct_global.R
\name{get_zeta_display_color_settings}
\alias{get_zeta_display_color_settings}
\title{Returns the BK colour setting}
\usage{
get_zeta_display_color_settings()
}
\value{
invisble return; function is called for side effects only
}
\description{
Returns the BK colour setting
}
