% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_BOTH_hypothesis_testing.R
\name{mod_hypothesis_testing_in}
\alias{mod_hypothesis_testing_in}
\title{Inputs related to data settings}
\usage{
mod_hypothesis_testing_in(id)
}
\arguments{
\item{id}{character to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a  \code{htmltools::\link[htmltools]{tagList}} container with
  \itemize{
     \item radio buttons to set time series frequency of displayed sales
     \item range input of weeks to use in control periods
     \item radio buttons to set the unit of the dependent variable (y-axis)
  }
}
\description{
Inputs related to data settings
}
