% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZetaDashboard.R
\docType{package}
\name{ZetaDashboard}
\alias{ZetaDashboard}
\title{ZetaDashboard}
\description{
A package for Hemskoep POS data for ZetaDisplaz
}
\section{Foo functions}{

The foo functions ...
}

\section{TO-DOS}{

\itemize{
\item improve documentation
\item write this package further
 }
}

